Move async_test to tests
