+++ tests/test_operators.py
@@ -8,7 +8,7 @@ import random
 import unittest
 
 from tests.mock_user import UserQuery
-from async_test import async_test
+from tests.async_test import async_test
 
 
 TEST_DATA = os.path.join(os.path.dirname(__file__), "test_data")
